manufacturing-ai-platform/services/agent-dispatcher/requirements.txt
fastapi
uvicorn
langchain
langgraph
redis
psycopg2-binary
requests
python-dotenv